requirements.txt
python_masscan==0.1.3
python_docx==0.8.10
requests==2.21.0
ipaddr==2.2.0
python_nmap==0.6.1
xlrd==1.2.0
pyfiglet==0.8.post1
ldap3==2.5.1
IPy==1.00
argparse==1.4.0
matplotlib